6.program/python/requirements.txt
# pip install -r requirements.txt

pywebview
flask
cx-Freeze